mod: model 변경
#S15P11D207-187

+++ frontend/modera/feature/category/src/main/java/com/ssafy/modera/feature/category/navigation/CategoryEntryProvider.kt
@@ -8,7 +8,7 @@ import com.ssafy.modera.feature.category.search.CategorySearchRoute
 fun EntryProviderScope<NavKey>.categoryEntry(
     onBackClick: () -> Unit,
     onSearchIconClick: () -> Unit,
-    onItemClick: (Int) -> Unit,
+    onItemClick: (Long) -> Unit,
 ) {
     entry<CategoryNavKey> {
         CategoryRoute(
